reworked predictions + integrated generator + refactor
The reworked predictions now add a padding around the image, and uses overlapped predictions (and discards some parts afterward) in order to avoid using the erroneous edge-predictions: this leaves the output image coming out much cleaner!
The generator is now integrated in the flow of the training, and I am now only waiting for to get back on the desktop to try to train a brand new model on the entire dataset (this laptop only works with CPU, which is too slow).
Refactored some code, removed a bunch of TODOs, cleaned some dead code (caused by the generator integration).
The readme was also updated to reflect some of the latests changes, but the actual results will be reupdated again to show how the fully-trained model performs.